Moved power-system data structures out of solve function
